refactor(web-serial-data-access): drop serial-command re-export (#673)
