--- libs/web-serial/data-access/src/lib/serial-command/serial-command-facade.service.ts
@@ -1,12 +1,0 @@
-/// <reference types="@types/w3c-web-serial" />
-
-/** Command API; 実装は {@link SerialCommandPipelineService}（#663）。 */
-export type {
-  CommandExecutionConfig,
-  CommandResult,
-} from './serial-command-types';
-
-export {
-  SerialCommandPipelineService,
-  type SerialCommandEnqueueOptions,
-} from './serial-command-pipeline.service';
